refactor: Add error handling in user routes

+++ backend/lib/backend_web/controllers/user_controller.ex
@@ -7,20 +7,32 @@ defmodule BackendWeb.UserController do
 
   action_fallback BackendWeb.FallbackController
 
-  # - a GET method : http://localhost:4000/api/users?email=XXX&username=YYY
-  # – a GET method : http://localhost:4000/api/users/:userID
-  # – a POST method : http://localhost:4000/api/users
-  # – a PUT method : http://localhost:4000/api/users/:userID
-  # – a DELETE method : http://localhost:4000/api/users/:userID
+  # GET method : http://localhost:4000/api/users?email=XXX&username=YYY
+  # GET method : http://localhost:4000/api/users/:userID
+  # POST method : http://localhost:4000/api/users
+  # PUT method : http://localhost:4000/api/users/:userID
+  # DELETE method : http://localhost:4000/api/users/:userID
 
   def index(conn, %{"email" => email, "username" => username}) do
     user = Accounts.get_user_by_username_and_email(username, email)
-    render(conn, :show, user: user)
+    case user do
+      nil -> conn
+      |> put_status(:not_found)
+      |> json(%{errors: ["User not found"]})
+      _ -> conn
+      render(conn, :show, user: user)
+    end
   end
 
   def show(conn, %{"id" => id}) do
-    user = Accounts.get_user!(id)
-    render(conn, :show, user: user)
+    user = Accounts.get_user(id)
+    case user do
+      nil -> conn
+      |> put_status(:not_found)
+      |> json(%{errors: ["User not found"]})
+      _ -> conn
+      render(conn, :show, user: user)
+    end
   end
 
   def create(conn, %{"user" => user_params}) do
@@ -33,18 +45,30 @@ defmodule BackendWeb.UserController do
   end
 
   def update(conn, %{"id" => id, "user" => user_params}) do
-    user = Accounts.get_user!(id)
+    user = Accounts.get_user(id)
 
-    with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
-      render(conn, :show, user: user)
+    if !user do
+      conn
+      |> put_status(:not_found)
+      |> json(%{errors: ["User not found, no update made"]})
+    else
+      with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
+        render(conn, :show, user: user)
+      end
     end
   end
 
   def delete(conn, %{"id" => id}) do
-    user = Accounts.get_user!(id)
+    user = Accounts.get_user(id)
 
-    with {:ok, %User{}} <- Accounts.delete_user(user) do
-      send_resp(conn, :no_content, "")
+    if !user do
+      conn
+      |> put_status(:not_found)
+      |> json(%{errors: ["User not found, no delete made"]})
+    else
+      with {:ok, %User{}} <- Accounts.delete_user(user) do
+        send_resp(conn, :no_content, "")
+      end
     end
   end
 end
